Move mutex to common
